Export default template
